backend/requirements.txt
Django==4.2.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.14.0
gunicorn==21.2.0
pandas
django-rest-knox
